requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# HTTP Requests
requests>=2.31.0
httpx>=0.25.2

# Environment Variables
python-dotenv>=1.0.0

# Image Processing (para detecção de duplicatas)
Pillow>=10.0.0
imagehash>=4.3.1

# PDF Processing
pdf2image>=1.17.0
pypdf2>=3.0.1
pdfplumber>=0.11.0

# OCR (opcional)
pytesseract>=0.3.10

# OpenAI (opcional, para OCR avançado)
openai>=1.0.0

# Date utilities
python-dateutil>=2.8.2
